Remove the duplicate definitions of link-resolver
