+++ common/views/components/PrismicHtmlBlock/PrismicHtmlBlock.tsx
@@ -1,4 +1,4 @@
-import { linkResolver } from '../../../services/prismic/link-resolver';
+import linkResolver from '@weco/common/services/prismic/link-resolver';
 import { JSXFunctionSerializer, PrismicRichText } from '@prismicio/react';
 import * as prismicT from '@prismicio/types';
 import { FC } from 'react';
