+++ content/webapp/types/card.ts
@@ -5,7 +5,7 @@ import { Article } from './articles';
 import { Season } from './seasons';
 import { Page, ParentPage } from './pages';
 import { Series } from './series';
-import linkResolver from '../services/prismic/link-resolver';
+import linkResolver from '@weco/common/services/prismic/link-resolver';
 import { EventSeries } from './event-series';
 import { Book } from './books';
 import { Exhibition } from './exhibitions';
